remove minor error causing bugs

+++ gate/orchestration/trainers/classification.py
@@ -52,7 +52,7 @@ class ClassificationTrainer(Trainer):
         for key, value in output_dict.items():
             if isinstance(value, torch.Tensor):
                 self.current_epoch_dict[key].append(
-                    value.detach().mean().cpu()
+                    value.detach().float().mean().cpu()
                 )
 
         return StepOutput(
